pyproject.toml
[tool.poetry]
name = "oidc"
version = "0.1.0"
description = ""
authors = ["Chance <chance@chancepage.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0"
django-oidc-provider = "^0.9.0"
requests = "^2.32.5"
django-two-factor-auth = "^1.18.1"
qrcode = "^8.2"
pyotp = "^2.9.0"
pillow = "^12.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
